Change library used for range calculations (again)
Since the last time I worked on this library, the `range` library has
been massively improved to basically be in line with my work on
`ord-ranges` as a library. Luckily, I never bothered to publish mine,
becuase it is now essentially obsolete and `range` has all the necessary
features required here.

However, the new range version is not available on the current Stackage
resolver.