Move to effect handling for Regex
Effect handling is done through Polysemy.

Schemas are now paramaterised by the type of the regex representation.
Both parsing and validating are also paramaterised by this type, and
effects exist for parsing into the representation and using the
representation to check when a string matches the regex.

Parser warnings and errors also now use an effect handler.